Actualización dia 22-03-2021

+++ web_widgets/customButtonInstanciarCatApyKey/customButtonInstanciarCatApyKey.ctrl.js
@@ -3,6 +3,7 @@ function PbButtonCtrl($scope, $http, $location, $log, $window, localStorageServi
     'use strict';
 
     var vm = this;
+    let claveValida = false;
 
     this.action = function action() {
         if ($scope.properties.action === 'Remove from collection') {
@@ -78,20 +79,29 @@ function startProcess() {
     if ($scope.properties.dataToChange2.campus || $scope.properties.dataToChange2.lstCatApiKeyInput[0].campus) {
         validar = true
         if ($scope.properties.dataToChange2.mailgun || $scope.properties.dataToChange2.mailgun === "") {
-            if ($scope.properties.dataToChange2.mailgun && $scope.properties.dataToChange2.mailgunCorreo && $scope.properties.dataToChange2.mailgunDominio && $scope.properties.dataToChange2.conekta && $scope.properties.dataToChange2.conektaPublicKey && $scope.properties.dataToChange2.crispChat) {
-                if ($scope.properties.processId) {
-                    var prom = doRequest('POST', '../API/bpm/process/' + $scope.properties.processId + '/instantiation', $scope.properties.userId).then(function () {
-                        doRequest("GET", $scope.properties.url).then(function () {
-                            $scope.properties.dataToChange = $scope.properties.dataToSet;
-                            $scope.properties.dataToChange2 = $scope.properties.dataToSet2;
+            if ($scope.properties.dataToChange2.mailgun && $scope.properties.dataToChange2.mailgunCorreo && $scope.properties.dataToChange2.mailgunDominio && $scope.properties.dataToChange2.conekta && $scope.properties.dataToChange2.conektaPublicKey && $scope.properties.dataToChange2.crispChat && $scope.properties.dataToChange2.hubspotKey) {
+                claveValida = validarClaveEditar($scope.properties.dataToChange2);
+                if(claveValida){
+                    if ($scope.properties.processId) {
+                        var prom = doRequest('POST', '../API/bpm/process/' + $scope.properties.processId + '/instantiation', $scope.properties.userId).then(function () {
+                            doRequest("GET", $scope.properties.url).then(function () {
+                                $scope.properties.dataToChange = $scope.properties.dataToSet;
+                                $scope.properties.dataToChange2 = $scope.properties.dataToSet2;
+                            });
+                            localStorageService.delete($window.location.href);
                         });
-                        localStorageService.delete($window.location.href);
-                    });
-
+    
+                    } else {
+                        $log.log('Impossible to retrieve the process definition id value from the URL');
+                    }
                 } else {
-                    $log.log('Impossible to retrieve the process definition id value from the URL');
+                    swal("¡Aviso!", "Ya están  registrados las API Keys del campus seleccionado", "warning");
                 }
+                
             } else {
+                if (!$scope.properties.dataToChange2.hubspotKey) {
+                    swal("¡Falto capurar informacion en!", "Hubspot key", "warning");
+                }
                 if (!$scope.properties.dataToChange2.crispChat) {
                     swal("¡Falto capurar informacion en!", "Crisp Chat key", "warning");
                 }
@@ -113,20 +123,29 @@ function startProcess() {
             }
 
         } else {
-            if ($scope.properties.dataToChange2.lstCatApiKeyInput[0].mailgun && $scope.properties.dataToChange2.lstCatApiKeyInput[0].mailgunCorreo && $scope.properties.dataToChange2.lstCatApiKeyInput[0].mailgunDominio && $scope.properties.dataToChange2.lstCatApiKeyInput[0].conekta && $scope.properties.dataToChange2.lstCatApiKeyInput[0].conektaPublicKey && $scope.properties.dataToChange2.lstCatApiKeyInput[0].crispChat) {
-                if ($scope.properties.processId) {
-                    var prom = doRequest('POST', '../API/bpm/process/' + $scope.properties.processId + '/instantiation', $scope.properties.userId).then(function () {
-                        doRequest("GET", $scope.properties.url).then(function () {
-                            $scope.properties.dataToChange = $scope.properties.dataToSet;
-                            $scope.properties.dataToChange2 = $scope.properties.dataToSet2;
+            if ($scope.properties.dataToChange2.lstCatApiKeyInput[0].mailgun && $scope.properties.dataToChange2.lstCatApiKeyInput[0].mailgunCorreo && $scope.properties.dataToChange2.lstCatApiKeyInput[0].mailgunDominio && $scope.properties.dataToChange2.lstCatApiKeyInput[0].conekta && $scope.properties.dataToChange2.lstCatApiKeyInput[0].conektaPublicKey && $scope.properties.dataToChange2.lstCatApiKeyInput[0].crispChat && $scope.properties.dataToChange2.lstCatApiKeyInput[0].hubspotKey) {
+                claveValida = validarNuevaClave($scope.properties.dataToChange2);
+                if(claveValida){
+                    if ($scope.properties.processId) {
+                        var prom = doRequest('POST', '../API/bpm/process/' + $scope.properties.processId + '/instantiation', $scope.properties.userId).then(function () {
+                            doRequest("GET", $scope.properties.url).then(function () {
+                                $scope.properties.dataToChange = $scope.properties.dataToSet;
+                                $scope.properties.dataToChange2 = $scope.properties.dataToSet2;
+                            });
+                            localStorageService.delete($window.location.href);
                         });
-                        localStorageService.delete($window.location.href);
-                    });
-
+    
+                    } else {
+                        $log.log('Impossible to retrieve the process definition id value from the URL');
+                    }
                 } else {
-                    $log.log('Impossible to retrieve the process definition id value from the URL');
+                    swal("¡Aviso!", "Ya están  registrados las API Keys del campus seleccionado", "warning");
                 }
+                
             } else {
+                if (!$scope.properties.dataToChange2.lstCatApiKeyInput[0].hubspotKey) {
+                    swal("¡Falto capurar informacion en!", "Hubspot key", "warning");
+                }
                 if (!$scope.properties.dataToChange2.lstCatApiKeyInput[0].crispChat) {
                     swal("¡Falto capurar informacion en!", "Crisp Chat key", "warning");
                 }
@@ -240,5 +259,35 @@ function startProcess() {
             $log.log('Impossible to retrieve the task id value from the URL');
         }
     }
+
+    function validarNuevaClave(_value){
+        let data = angular.copy($scope.properties.dataFromSuccess);
+        let isValid = true;
+        
+        for(let i = 0; i< data.length; i++){
+            if(data[i].campus.descripcion.toLowerCase() === _value.lstCatApiKeyInput[0].campus.descripcion.toLowerCase()){
+                isValid = false;
+                break;
+            }
+        }
+    
+        return isValid;
+    }
+    
+    function validarClaveEditar(_value){
+        let data = angular.copy($scope.properties.dataFromSuccess);
+        let isValid = true;
+        let pidString = _value.persistenceId + "";
+        
+        for(let i = 0; i< data.length; i++){
+            if(data[i].campus.descripcion.toLowerCase() === _value.campus.descripcion.toLowerCase() && pidString !== (data[i].persistenceId + "")){
+                isValid = false;
+                break;
+            }
+        }
+    
+        return isValid;
+    }
+    
 
 }
